Move the preallocate tests into their own files
And move the MetaAddr proptest into its own file.

Also do some minor formatting and cleanups.
