+++ zebra-chain/src/block/tests.rs
@@ -1,4 +1,5 @@
-// XXX this should be rewritten as strategies
+// XXX generate should be rewritten as strategies
 mod generate;
+mod preallocate;
 mod prop;
 mod vectors;
