+++ zebra-chain/src/transaction/tests.rs
@@ -1,2 +1,3 @@
+mod preallocate;
 mod prop;
 mod vectors;
