+++ zebra-network/src/meta_addr.rs
@@ -18,6 +18,9 @@ use crate::protocol::{external::MAX_PROTOCOL_MESSAGE_LEN, types::PeerServices};
 
 use PeerAddrState::*;
 
+#[cfg(test)]
+mod tests;
+
 /// Peer connection state, based on our interactions with the peer.
 ///
 /// Zebra also tracks how recently a peer has sent us messages, and derives peer
@@ -264,105 +267,14 @@ impl ZcashDeserialize for MetaAddr {
         Ok(MetaAddr::new_gossiped(&addr, &services, &last_seen))
     }
 }
+
 /// A serialized meta addr has a 4 byte time, 8 byte services, 16 byte IP addr, and 2 byte port
 const META_ADDR_SIZE: usize = 4 + 8 + 16 + 2;
+
 impl TrustedPreallocate for MetaAddr {
     fn max_allocation() -> u64 {
         // Since a maximal serialized Vec<MetAddr> uses at least three bytes for its length (2MB  messages / 30B MetaAddr implies the maximal length is much greater than 253)
         // the max allocation can never exceed (MAX_PROTOCOL_MESSAGE_LEN - 3) / META_ADDR_SIZE
         ((MAX_PROTOCOL_MESSAGE_LEN - 3) / META_ADDR_SIZE) as u64
-    }
-}
-
-#[cfg(test)]
-mod test_trusted_preallocate {
-    use std::convert::TryInto;
-
-    use super::{MetaAddr, MAX_PROTOCOL_MESSAGE_LEN, META_ADDR_SIZE};
-    use super::{PeerAddrState, PeerServices};
-    use chrono::{TimeZone, Utc};
-    use zebra_chain::serialization::{TrustedPreallocate, ZcashSerialize};
-    #[test]
-    /// Confirm that each MetaAddr takes exactly META_ADDR_SIZE bytes when serialized.
-    /// This verifies that our calculated `TrustedPreallocate::max_allocation()` is indeed an upper bound.
-    fn meta_addr_size_is_correct() {
-        let addr = MetaAddr {
-            addr: ([192, 168, 0, 0], 8333).into(),
-            services: PeerServices::default(),
-            last_seen: Utc.timestamp(1_573_680_222, 0),
-            last_connection_state: PeerAddrState::Responded,
-        };
-        let serialized = addr
-            .zcash_serialize_to_vec()
-            .expect("Serialization to vec must succeed");
-        assert!(serialized.len() == META_ADDR_SIZE)
-    }
-    #[test]
-    /// Verifies that...
-    /// 1. The smallest disallowed vector of `MetaAddrs`s is too large to fit in a legal Zcash message
-    /// 2. The largest allowed vector is small enough to fit in a legal Zcash message
-    fn meta_addr_max_allocation_is_correct() {
-        let addr = MetaAddr {
-            addr: ([192, 168, 0, 0], 8333).into(),
-            services: PeerServices::default(),
-            last_seen: Utc.timestamp(1_573_680_222, 0),
-            last_connection_state: PeerAddrState::Responded,
-        };
-        let max_allocation: usize = MetaAddr::max_allocation().try_into().unwrap();
-        let mut smallest_disallowed_vec = Vec::with_capacity(max_allocation + 1);
-        for _ in 0..(MetaAddr::max_allocation() + 1) {
-            smallest_disallowed_vec.push(addr);
-        }
-        let smallest_disallowed_serialized = smallest_disallowed_vec
-            .zcash_serialize_to_vec()
-            .expect("Serialization to vec must succeed");
-        // Check that our smallest_disallowed_vec is only one item larger than the limit
-        assert!(((smallest_disallowed_vec.len() - 1) as u64) == MetaAddr::max_allocation());
-        // Check that our smallest_disallowed_vec is too big to send in a valid Zcash message
-        assert!(smallest_disallowed_serialized.len() > MAX_PROTOCOL_MESSAGE_LEN);
-
-        // Create largest_allowed_vec by removing one element from smallest_disallowed_vec without copying (for efficiency)
-        smallest_disallowed_vec.pop();
-        let largest_allowed_vec = smallest_disallowed_vec;
-        let largest_allowed_serialized = largest_allowed_vec
-            .zcash_serialize_to_vec()
-            .expect("Serialization to vec must succeed");
-
-        // Check that our largest_allowed_vec contains the maximum number of MetaAddrs
-        assert!((largest_allowed_vec.len() as u64) == MetaAddr::max_allocation());
-        // Check that our largest_allowed_vec is small enough to fit in a Zcash message.
-        assert!(largest_allowed_serialized.len() <= MAX_PROTOCOL_MESSAGE_LEN);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    // XXX remove this test and replace it with a proptest instance.
-    #[test]
-    fn sanitize_truncates_timestamps() {
-        zebra_test::init();
-
-        let services = PeerServices::default();
-        let addr = "127.0.0.1:8233".parse().unwrap();
-
-        let entry = MetaAddr {
-            services,
-            addr,
-            last_seen: Utc.timestamp(1_573_680_222, 0),
-            last_connection_state: Responded,
-        }
-        .sanitize();
-
-        // We want the sanitized timestamp to be a multiple of the truncation interval.
-        assert_eq!(
-            entry.get_last_seen().timestamp() % crate::constants::TIMESTAMP_TRUNCATION_SECONDS,
-            0
-        );
-        // We want the state to be the default
-        assert_eq!(entry.last_connection_state, Default::default());
-        // We want the other fields to be unmodified
-        assert_eq!(entry.addr, addr);
-        assert_eq!(entry.services, services);
     }
 }
